[KYUUBI #6975] Clean up code for Spark 3.5 extension
### Why are the changes needed?

Simple refactoring to clean up the code for the Spark 3.5 extension, e.g., remove unnecessary `*Base` `*Helper` abstraction layers, remove code for legacy Spark versions.

Note: I don't touch `ForcedMaxOutputRows*` because I'm going to remove it in the next PR.

Preparation for Spark 4.0 support.

### How was this patch tested?

Pass GHA.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #6975 from pan3793/spark-ext-35-cleanup.

Closes #6975

b5a94a680 [Cheng Pan] nit
c729e268c [Cheng Pan] fix
1087ac709 [Cheng Pan] Clean up code for Spark 3.5 extension

Authored-by: Cheng Pan <chengpan@apache.org>
