+++ extensions/spark/kyuubi-extension-spark-3-5/src/main/scala/org/apache/kyuubi/sql/KyuubiSparkSQLAstBuilder.scala
@@ -179,7 +179,7 @@ trait UnparsedExpressionLogicalPlan extends LogicalPlan {
 
   override def children: Seq[LogicalPlan] = throw new UnsupportedOperationException()
 
-  protected def withNewChildrenInternal(
+  override def withNewChildrenInternal(
       newChildren: IndexedSeq[LogicalPlan]): LogicalPlan =
     throw new UnsupportedOperationException()
 }
